refactor(frontend): the project screen has one story file, and the capped files stay inside the ratchet
The three ProjectScreen stories move beside the screen with the folded-pane
state and share one route stub, so the catalog has one entry per screen. The
deals route, the shell's record test and the column comments shrink to the
lengths the file-length waivers freeze.
